+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2018-12-04
+### Added
+
 ## [0.2.3] - 2018-12-04
 ### Added
 - `SQLiteDatabase#delete()` method
@@ -98,6 +101,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SQLiteDatabase#rawQuery()` method
 - `SQLiteDatabase#setLocale()` method
 
+[0.2.4]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.3...0.2.4
 [0.2.3]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.2...0.2.3
 [0.2.2]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.1...0.2.2
 [0.2.1]:  https://github.com/drydart/flutter_sqlcipher/compare/0.2.0...0.2.1